feat: create sole summary data handler, moved existing logic, added new by language filtering
